+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## 1.0
 
 ### Unreleased
+- NEW : Update total cost of bom with general expenses extrafield - *01/10/2025* - 1.2.0
 - NEW : add Extrafield thirdparty ref required(boolean) use on trigger order_validate & massaction *24/09/2025* - 1.2.0
 - NEW : Surface calculation from length and height on order/propal lines *24/09/2025* - 1.2.0
 - NEW : automated contract tariff revision feature that updates contracts line prices on a scheduled date and notifies relevant users of the change  - 1.1.0